# docs/gitbook/api/openapi.yaml
openapi: 3.1.0
info:
  description: API Server for Cyphera application
  title: Cyphera API
  termsOfService: http://swagger.io/terms/
  contact:
    name: API Support
    url: http://www.swagger.io/support
    email: support@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: "1.0"
paths:
...
  /wallets:
    get:
      security:
        - ApiKeyAuth: []
      description: List all wallets for the authenticated workspace
      tags:
        - wallets
      summary: List all wallets
      parameters:
        - description: Include Circle wallet data
          name: include_circle_data
          in: query
          schema:
            type: boolean
        - description: Filter by wallet type (wallet, circle_wallet)
          name: wallet_type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.WalletListResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.ErrorResponse"
    post:
      security:
        - ApiKeyAuth: []
      description: Creates a new wallet for each active network in the authenticated
        workspace
      tags:
        - wallets
      summary: Create a new wallet
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/handlers.CreateWalletRequest"
        description: Wallet creation request
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.WalletListResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.ErrorResponse"
  "/wallets/{wallet_id}":
    get:
      security:
        - ApiKeyAuth: []
      description: Get wallet details by wallet ID
      tags:
        - wallets
      summary: Get wallet by ID
      parameters:
        - description: Wallet ID
          name: wallet_id
          in: path
          required: true
          schema:
            type: string
        - description: Include Circle wallet data
          name: include_circle_data
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.WalletResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.ErrorResponse"
    delete:
      security:
        - ApiKeyAuth: []
      description: Soft delete a wallet
      tags:
        - wallets
      summary: Delete a wallet
      parameters:
        - description: Wallet ID
          name: wallet_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.ErrorResponse"
...
servers:
  - url: //localhost:8000/api/v1
components:
  securitySchemes:
    ApiKeyAuth:
      description: API Key authentication. Use 'Bearer <token>' format.
      type: apiKey
      name: Authorization
      in: header
    Bearer:
      description: Type "Bearer" followed by a space and JWT token.
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    handlers.CircleWalletData:
      type: object
      properties:
        chain_id:
          type: integer
        circle_user_id:
          type: string
        circle_wallet_id:
          type: string
        state:
          type: string
...
    handlers.CreateWalletRequest:
      type: object
      required:
        - network_type
        - wallet_address
        - wallet_type
      properties:
        chain_id:
          description: Only for circle wallets
          type: integer
        circle_user_id:
          description: Circle wallet specific fields
          type: string
        circle_wallet_id:
          description: Only for circle wallets
          type: string
        ens:
          type: string
        is_primary:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
        network_type:
          description: "'evm' or 'solana'"
          type: string
        nickname:
          type: string
        state:
          description: Only for circle wallets
          type: string
        verified:
          type: boolean
        wallet_address:
          type: string
        wallet_type:
          description: "'wallet' or 'circle_wallet'"
          type: string
...
    handlers.ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    handlers.WalletListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/handlers.WalletResponse"
        object:
          type: string
    handlers.WalletResponse:
      type: object
      properties:
        circle_data:
          description: Only present for circle wallets
          allOf:
            - $ref: "#/components/schemas/handlers.CircleWalletData"
        created_at:
          type: integer
        ens:
          type: string
        id:
          type: string
        is_primary:
          type: boolean
        last_used_at:
          type: integer
        metadata:
          type: object
          additionalProperties: true
        network_id:
          type: string
        network_type:
          type: string
        nickname:
          type: string
        object:
          type: string
        updated_at:
          type: integer
        verified:
          type: boolean
        wallet_address:
          type: string
        wallet_type:
          description: "'wallet' or 'circle_wallet'"
          type: string
        workspace_id:
          type: string
...